# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perpustakaan",
    "description": "Perpustakaan API Collection",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3880/"
    }
  ],
  "paths": {
...
    "/koleksibuku": {
      "post": {
        "tags": [
          "Koleksibuku"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KoleksibukuAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Koleksibuku"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KoleksibukuUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Koleksibuku"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KoleksibukuDelete"
              }
            }
          }
        }
      }
    },
...
    "/koleksibuku/{id}": {
      "get": {
        "tags": [
          "Koleksibuku"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/transaksi": {
      "post": {
        "tags": [
          "Transaksi"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransaksiAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Transaksi"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransaksiUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transaksi"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransaksiDelete"
              }
            }
          }
        }
      }
    },
...
    "/transaksi/{id}": {
      "get": {
        "tags": [
          "Transaksi"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KoleksibukuAdd": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "koleksibuku label"
          }
        }
      },
      "KoleksibukuUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "koleksibuku label"
          },
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
      "KoleksibukuDelete": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
...
      "TransaksiAdd": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "transaksi label"
          }
        }
      },
      "TransaksiUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "transaksi label"
          },
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
      "TransaksiDelete": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 0
          }
        }
      }
    }
  }
}